factor setup code out of server.rs, break calls up in main.rs
